backpack_finder/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12
pillow==10.4.0
numpy==2.0.2

ultralytics==8.3.55

streamlit==1.41.1
requests==2.32.3

openpyxl==3.1.5
reportlab==4.2.5

torch==2.5.1
torchvision==0.20.1
